Add configuration for song age checking
This commit adds a new environment variable 'DAY_DIFF' to set the limit of how old a song can be to be considered relevant. In 'main.py', logic has been added to determine the release date of the current song. If the song is older than the 'DAY_DIFF' setting, it's considered too old, and processing is skipped. This change allows customization over song relevancy based on its age.